Move ReactiveTransportResult to its own file
